A new class was created to encapsulate two covid data files
